% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bngeneplotCustom.R
\name{bngeneplotCustom}
\alias{bngeneplotCustom}
\title{bngeneplotCustom}
\usage{
bngeneplotCustom(
  results,
  exp,
  expSample = NULL,
  algo = "hc",
  R = 20,
  pathNum = NULL,
  convertSymbol = TRUE,
  expRow = "ENSEMBL",
  interactive = FALSE,
  cexCategory = 1,
  cl = NULL,
  showDir = FALSE,
  chooseDir = FALSE,
  algorithm.args = NULL,
  labelSize = 4,
  layout = "nicely",
  strType = "normal",
  returnNet = FALSE,
  otherVar = NULL,
  otherVarName = NULL,
  onlyDf = FALSE,
  disc = FALSE,
  tr = NULL,
  remainCont = NULL,
  dep = NULL,
  sizeDep = FALSE,
  orgDb = org.Hs.eg.db,
  cellLineName = "5637_URINARY_TRACT",
  fontFamily = "sans",
  showLineage = FALSE,
  strengthPlot = FALSE,
  nStrength = 10,
  strThresh = NULL,
  hub = NULL,
  glowEdgeNum = NULL,
  nodePal = c("blue", "red"),
  edgePal = c("blue", "red"),
  textCol = "black",
  titleCol = "black",
  backCol = "white",
  barTextCol = "black",
  barPal = c("red", "blue"),
  barBackCol = "white",
  scoreType = "bic-g",
  barLegendKeyCol = "white",
  barAxisCol = "black",
  barPanelGridCol = "black",
  titleSize = 24
)
}
\arguments{
\item{results}{the enrichment analysis result from clusterProfiler or ReactomePA}

\item{exp}{gene expression matrix}

\item{expSample}{candidate rows to be included in the inference, default to all}

\item{algo}{structure learning method used in boot.strength(), default to "hc"}

\item{R}{the number of bootstrap}

\item{pathNum}{the pathway number (the number of row of the original result, ordered by p-value)}

\item{convertSymbol}{whether the label of resulting network is converted to symbol, default to TRUE}

\item{expRow}{the type of the identifier of rows of expression matrix}

\item{cexCategory}{scaling factor of size of nodes}

\item{cl}{cluster object from parallel::makeCluster()}

\item{showDir}{show the confidence of direction of edges}

\item{chooseDir}{if undirected edges are present, choose direction of edges}

\item{labelSize}{the size of label of the nodes}

\item{layout}{ggraph layout, default to "nicely"}

\item{strType}{"normal" or "ms" for multiscale implementation}

\item{returnNet}{whether to return the network}

\item{dep}{the tibble storing dependency score from library depmap}

\item{sizeDep}{whether to reflect DepMap score to the node size}

\item{cellLineName}{the cell line name to be included}

\item{fontFamily}{font family name to be used for plotting}

\item{showLineage}{show the dependency score across the lineage}

\item{strengthPlot}{append the barplot depicting edges with high strength}

\item{nStrength}{specify how many edges are included in the strength plot}

\item{strThresh}{the threshold for strength}

\item{hub}{visualize the genes with top-n hub scores}

\item{glowEdgeNum}{edges with top-n confidence of direction are highlighted}

\item{nodePal}{vector of coloring of nodes (low, high)}

\item{edgePal}{vector of coloring of edges (low, high)}

\item{textCol}{color of texts in network plot}

\item{titleCol}{color of title in network plot}

\item{backCol}{color of background in network plot}

\item{barTextCol}{text color in barplot}

\item{barPal}{bar color}

\item{barBackCol}{background color in barplot}

\item{scoreType}{score type to use on inference}

\item{barLegendKeyCol}{legend key color in barplot}

\item{barAxisCol}{axis color in barplot}

\item{barPanelGridCol}{panel grid color in barplot}

\item{titleSize}{the size of title}

\item{pathDb}{query to graphite::pathways(), default to "reactome"}

\item{depMeta}{depmap::depmap_metadata(), needed for showLineage}
}
\value{
ggplot2 object
}
\description{
Plot gene relationship within the specified pathway using customized theme
}
\examples{
bngeneplotCustom(results = pway, chooseDir=TRUE, pathNum=15, glowEdgeNum=NULL, hub=3, R=40, fontFamily="sans", exp = vsted, expSample=rownames(subset(meta, Condition=="T")))
}
