% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bootReasonMultiple}
\alias{bootReasonMultiple}
\title{bootReasonMultiple}
\usage{
bootReasonMultiple(
  df,
  R,
  nodes,
  evidences,
  levels,
  algo = "hc",
  n = NULL,
  algorithm.args = NULL
)
}
\arguments{
\item{df}{data frame to perform structure learning}

\item{R}{bootstrap replicates number}

\item{nodes}{node names of the event}

\item{evidences}{node names of the evidence}

\item{algo}{structure learning algorithm, default to hc}

\item{algorithm.args}{passed to algorithm function}

\item{level}{evidence levels}
}
\value{
the mean of the bootstrapped values for each level
}
\description{
bootstrapping probabilistic reasoning and return the mean of bootstrapped values.
Multiple events, evidences and levels can be specified. The likelihood weighting was used for sampling.
}
\examples{
bootReasonMultiple(df, R, nodes=c("A"), evidences=c("B","C"), levels=c(2,3))
}
