% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBNplot-package.R
\docType{package}
\name{CBNplot-package}
\alias{CBNplot}
\alias{CBNplot-package}
\title{CBNplot: plot bayesian network inferred from gene expression data based on enrichment analysis results}
\description{
This package provides the visualization of bayesian network inferred from gene expression data. The networks are based on enrichment analysis results inferred from packages including clusterProfiler and ReactomePA. The networks between pathways and genes inside the pathways can be inferred and visualized.
}
\details{
[CBNplot::bngeneplot()] - The main function using gene expression data
                          within the pathway to infer Bayesian network.
[CBNplot::bnpathplot()] - The main function using pathway expression data,
                          which is defined by eigengene of the gene
                          expression values to infer Bayesian network.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noriakis/CBNplot}
  \item Report bugs at \url{https://github.com/noriakis/CBNplot/issues}
}

}
\author{
\strong{Maintainer}: Noriaki Sato \email{nori5030@kuhp.kyoto-u.ac.jp}

}
\keyword{internal}
