% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compareBNs}
\alias{compareBNs}
\title{compareBNs}
\usage{
compareBNs(listOfNets)
}
\arguments{
\item{listOfNets}{list of networks}
}
\value{
F-measures of each combination of network
}
\description{
Take the list of networks and returns the F-measures
}
\examples{
data("exampleEaRes");data("exampleGeneExp")
net1 <- bngeneplot(results = exampleEaRes, exp = exampleGeneExp, pathNum = 1, R = 10, returnNet=TRUE)
net2 <- bngeneplot(results = exampleEaRes, exp = exampleGeneExp, pathNum = 1, R = 10, returnNet=TRUE)
res <- compareBNs(list(net1$av, net2$av))
}
