% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{depKStest}
\alias{depKStest}
\title{depKStest}
\usage{
depKStest(
  results,
  type,
  cellLineName = NULL,
  lineageName = NULL,
  adjMethod = "bonferroni",
  dep = NULL,
  depMeta = NULL
)
}
\arguments{
\item{type}{"cell_line" or "lineage"}
}
\value{
pvalues and adjusted pvalues
}
\description{
Perform KS-test recursively
}
\examples{
depKStest(pway, type="cell_line", cellLineName="253J_URINARY_TRACT", dep=depmap::depmap_crispr(), depMeta=depmap::depmap_metadata())
}
