% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleEaRes}
\alias{exampleEaRes}
\title{Example enrichment analysis result}
\format{
An object of class \code{enrichResult} with 47 rows and 9 columns.
}
\usage{
exampleEaRes
}
\description{
An example enrichment analysis result to be used for testing purpose
}
\keyword{datasets}
