% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleGeneExp}
\alias{exampleGeneExp}
\title{Example gene expression data}
\format{
An object of class \code{data.frame} with 7 rows and 100 columns.
}
\usage{
data(exampleGeneExp)
}
\value{
example gene expression
}
\description{
An example gene expression data to be used for testing purpose
made by runif() for ERCC genes and 100 samples.
No biological meanings can be obtained from the data.
}
\keyword{datasets}
