% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inferMS}
\alias{inferMS}
\title{inferMS}
\usage{
inferMS(data, algo, algorithm.args, R, cl = NULL, r = seq(0.5, 1.5, 0.1))
}
\value{
object of class bn.strength
}
\description{
multiscale bootstrap-based inference of Bayesian network
}
