% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadSign}
\alias{loadSign}
\title{loadSign}
\usage{
loadSign(fileName)
}
\arguments{
\item{fileName}{the result of SiGN-BN}
}
\value{
list of edges, nodes, strength, and bn (bnlearn)
}
\description{
Load the output of SiGN-BN (HC+BS)
}
