% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{obtainPath}
\alias{obtainPath}
\title{obtainPath}
\usage{
obtainPath(res, geneSymbol)
}
\arguments{
\item{res}{enrichment analysis result}

\item{geneSymbol}{the candidate gene}
}
\value{
subset of enrichment results
}
\description{
obtain the analysis results including the queried gene symbol
}
\examples{
obtainPath(results = pway, geneSymbol="E2F1")
}
