% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bngeneplot.R
\name{bngeneplot}
\alias{bngeneplot}
\title{bngeneplot}
\usage{
bngeneplot(
  results,
  exp,
  expSample = NULL,
  algo = "hc",
  R = 20,
  returnNet = FALSE,
  algorithm.args = NULL,
  pathNum = NULL,
  convertSymbol = TRUE,
  expRow = "ENSEMBL",
  interactive = FALSE,
  cexCategory = 1,
  cl = NULL,
  showDir = FALSE,
  chooseDir = FALSE,
  scoreType = "bic-g",
  labelSize = 4,
  layout = "nicely",
  clusterAlpha = 0.2,
  strType = "normal",
  delZeroDegree = TRUE,
  otherVar = NULL,
  otherVarName = NULL,
  onlyDf = FALSE,
  disc = FALSE,
  tr = NULL,
  remainCont = NULL,
  sp = "hsapiens",
  compareRef = FALSE,
  compareRefType = "intersection",
  pathDb = "reactome",
  dep = NULL,
  depMeta = NULL,
  sizeDep = FALSE,
  showDepHist = TRUE,
  cellLineName = "5637_URINARY_TRACT",
  showLineage = FALSE,
  orgDb = org.Hs.eg.db,
  shadowText = FALSE,
  strengthPlot = FALSE,
  nStrength = 10,
  strThresh = NULL,
  hub = NULL
)
}
\arguments{
\item{results}{the enrichment analysis result from clusterProfiler or ReactomePA}

\item{exp}{gene expression matrix}

\item{expSample}{candidate samples to be included in the inference, default to all}

\item{algo}{structure learning method used in boot.strength(), default to "hc"}

\item{R}{the number of bootstrap}

\item{returnNet}{whether to return the network}

\item{pathNum}{the pathway number (the number of row of the original result, ordered by p-value)}

\item{convertSymbol}{whether the label of resulting network is converted to symbol, default to TRUE}

\item{expRow}{the type of the identifier of rows of expression matrix}

\item{interactive}{whether to use bnviewer (default to FALSE)}

\item{cexCategory}{scaling factor of size of nodes}

\item{cl}{cluster object from parallel::makeCluster()}

\item{showDir}{show the confidence of direction of edges}

\item{chooseDir}{if undirected edges are present, choose direction of edges (default: FALSE)}

\item{scoreType}{score type to use on choosing direction}

\item{labelSize}{the size of label of the nodes}

\item{layout}{ggraph layout, default to "nicely"}

\item{clusterAlpha}{if specified multiple pathways, the parameter is passed to geom_mark_hull()}

\item{strType}{"normal" or "ms" for multiscale implementation}

\item{otherVar}{other variables to be included in the inference}

\item{otherVarName}{the names of other variables}

\item{onlyDf}{return only data.frame used for inference}

\item{sp}{query to graphite::pathways(), default to "hsapiens"}

\item{compareRef}{whether compare to the reference network}

\item{compareRefType}{"intersection" or "difference"}

\item{pathDb}{query to graphite::pathways(), default to "reactome"}

\item{dep}{the tibble storing dependency score from library depmap}

\item{depMeta}{depmap::depmap_metadata(), needed for showLineage}

\item{sizeDep}{whether to reflect DepMap score to the node size}

\item{cellLineName}{the cell line name to be included}

\item{showLineage}{show the dependency score across the lineage}

\item{strengthPlot}{append the barplot depicting edges with high strength}

\item{nStrength}{specify how many edges are included in the strength plot}

\item{strThresh}{the threshold for strength}

\item{hub}{visualize the genes with top-n hub scores}
}
\value{
ggplot2 object
}
\description{
Plot gene relationship within the specified pathway
}
\examples{
bngeneplot(results = pway, exp = vsted, expSample = rownames(subset(meta, Condition=="T")), pathNum = 17, R = 10, convertSymbol = TRUE, expRow = "ENSEMBL")

}
