% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compareBNs}
\alias{compareBNs}
\title{compareBNs}
\usage{
compareBNs(listOfNets)
}
\arguments{
\item{nets}{data frame to perform structure learning}
}
\value{
F-measures of each combination of network
}
\description{
Take the list of networks and returns the F-measures
}
\examples{
compareBNs(nets)
}
