% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inferMS}
\alias{inferMS}
\title{inferMS}
\usage{
inferMS(data, algo, algorithm.args, R, cl = NULL, r = seq(0.5, 1.5, 0.1))
}
\arguments{
\item{data}{data.frame to perform inference}

\item{algo}{structure learning method used in boot.strength()}

\item{algorithm.args}{parameters to pass to
bnlearn structure learnng function}

\item{R}{the number of bootstrap}

\item{cl}{cluster object from parallel::makeCluster()}

\item{r}{vector for size of each bootstrap replicate}
}
\value{
object of class bn.strength
}
\description{
multiscale bootstrap-based inference of Bayesian network
}
