% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{queryCpDistLw}
\alias{queryCpDistLw}
\title{queryCpDistLw}
\usage{
queryCpDistLw(
  fitted,
  candidate,
  evidence,
  level,
  point = FALSE,
  pointSize = 5,
  alpha = TRUE,
  ...
)
}
\arguments{
\item{fitted}{bn.fit object}

\item{candidate}{name of node}

\item{evidence}{evidence variable name}

\item{level}{level to be listed}

\item{point}{geom_point the weighted mean}

\item{alpha}{whether to reflect the weights by alpha (TRUE) or color (FALSE)}
}
\value{
list of dataframe containing raw values
}
\description{
produce a plot of bnlearn::cpdist by performing bnlearn::cpdist on specified node, evidence and level.
}
\examples{
queryCpDistLw(fitted, candidate="COL25A1", evidence = "Treatment", level=rownames(fitted$Treatment$prob))
}
